backend/requirements.txt
fastapi==0.103.1          # 轻量高效的后端框架
uvicorn==0.23.2           # 运行FastAPI的ASGI服务器
pydantic==2.4.2           # 数据校验
pydantic-settings==2.0.3  # 配置文件管理
python-dotenv==1.0.0      # 加载.env环境变量
openai==1.3.5             # OpenAI API调用（可选，辅助解析）
python-multipart==0.0.6   # 处理表单数据（备用）